Changed packages structure. Moved Book service related to book package.
